+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ### [TURISTEA-BE]
+### [0.0.4] - 18-10-2025
+-feat: role profile and create user driver with validate
+
+### [0.0.3] - 18-10-2025
+-feat: Upload Readme and delete comments
+-feat: endpoint admin and profile client
+-feat: implements model package and drivers
+-feat: endpoint driver for administrator
+-feat: add cellphone for profile client
+-feat: implements photo for drivers
 
 ### [0.0.2] - 04-08-2025
 ### Feature